Address review comments
* Fix trivial var naming and refactoring
* Delete un-related header including
* Add BraveTorClientUpdater::Cleanup() instead of deleting tor extension
  directly from BraveExtensionManagement.
