+++ browser/extensions/brave_tor_client_updater.cc
@@ -14,9 +14,18 @@
 #include "base/task/post_task.h"
 #include "base/task_runner.h"
 #include "base/task_runner_util.h"
+#include "brave/browser/tor/tor_profile_service.h"
+#include "chrome/browser/browser_process.h"
+#include "chrome/browser/profiles/profile_manager.h"
 #include "third_party/re2/src/re2/re2.h"
 
 using brave_component_updater::BraveComponent;
+
+namespace {
+void DeleteDir(const base::FilePath& path) {
+  base::DeleteFile(path, true);
+}
+}  // namespace
 
 namespace extensions {
 
@@ -117,6 +126,17 @@ void BraveTorClientUpdater::Register() {
                                g_tor_client_component_id_,
                                g_tor_client_component_base64_public_key_);
   registered_ = true;
+}
+
+void BraveTorClientUpdater::Cleanup() {
+  // Delete tor binaries if tor is disabled by gpo.
+  if (tor::TorProfileService::IsTorDisabled()) {
+    ProfileManager* profile_manager = g_browser_process->profile_manager();
+    base::FilePath tor_component_dir =
+        profile_manager->user_data_dir().AppendASCII(kTorClientComponentId);
+    GetTaskRunner()->PostTask(FROM_HERE,
+                              base::BindOnce(&DeleteDir, tor_component_dir));
+  }
 }
 
 void BraveTorClientUpdater::SetExecutablePath(const base::FilePath& path) {
